Added esbuild option for testing
make build "lints" the js file making it impossible to add debugger statements into the source.
Changing jshint to ignore debugger statements would be a bad idea as they should never showup in
production, so the alternative is to use `make esbuilkd` when testing code, then just `make` when
you want a production ready version of the code.